Make the bot and router classes immutable
Since the bot and router classes themselves represent the concept of the
bot and the router, rather than instances of those classes (which may or
may not be madness), we must make them immutable for reasons of thread
safety. We can’t have threads changing the bot or router willy-nilly,
as that might step on other threads also attempting to change those
classes.